scripts/requirements.txt

mysql-connector-python
requests
beautifulsoup4
